requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
python-dotenv
httpx
coincurve
slowapi
pytest
pytest-asyncio
